15 februari 2023 10:23 : -minor fix update profile user

+++ app/Services/UserService.php
@@ -134,8 +134,10 @@ class UserService
                         isset($request['foto']),
                         function ($q) use ($request, $user) {
 
-                            if (Storage::exists($user->employee->foto)) {
-                                Storage::delete($user->employee->foto);
+                            if ($user->employee->foto != null) {
+                                if (Storage::exists($user->employee->foto)) {
+                                    Storage::delete($user->employee->foto);
+                                }
                             }
 
                             $file = $request['foto'];
